+++ conf/database.properties
@@ -22,6 +22,7 @@ scalar.db.password=cassandra
 # If SNAPSHOT is specified in the property "scalar.db.consensus_commit.isolation_level", this is ignored.
 #scalar.db.consensus_commit.serializable_strategy=
 
+# Debug mode usable for Consensus Commit.
 # If set to "true", Get and Scan operations results will contain transaction metadata. To see the transaction metadata columns details for a given table, you can use the `DistributedTransactionAdmin.getTableMetadata()`
 # method which will return the table metadata augmented with the transaction metadata columns. Using the debug mode can be useful to investigate transaction related issues.
 # The default is false.
