+++ core/src/integration-test/java/com/scalar/db/transaction/jdbc/JdbcTransactionAdminIntegrationTest.java
@@ -4,7 +4,6 @@ import com.scalar.db.api.DistributedTransactionAdminIntegrationTestBase;
 import com.scalar.db.config.DatabaseConfig;
 import com.scalar.db.storage.jdbc.JdbcEnv;
 import java.util.Properties;
-import org.junit.jupiter.api.Disabled;
 
 public class JdbcTransactionAdminIntegrationTest
     extends DistributedTransactionAdminIntegrationTestBase {
@@ -16,9 +15,4 @@ public class JdbcTransactionAdminIntegrationTest
     properties.setProperty(DatabaseConfig.TRANSACTION_MANAGER, "jdbc");
     return properties;
   }
-
-  @Disabled("There is no metadata columns when using the jdbc transaction manager")
-  @Override
-  public void
-      getTableMetadata_WhenDebugging_ShouldReturnCorrectMetadataWithTransactionMetadataColumns() {}
 }
