+++ core/src/integration-test/java/com/scalar/db/transaction/jdbc/JdbcTransactionIntegrationTest.java
@@ -2,11 +2,9 @@ package com.scalar.db.transaction.jdbc;
 
 import com.scalar.db.api.DistributedTransactionIntegrationTestBase;
 import com.scalar.db.config.DatabaseConfig;
-import com.scalar.db.exception.transaction.TransactionException;
 import com.scalar.db.storage.jdbc.JdbcEnv;
 import java.util.Properties;
 import org.junit.jupiter.api.Disabled;
-import org.junit.jupiter.api.Test;
 
 public class JdbcTransactionIntegrationTest extends DistributedTransactionIntegrationTestBase {
 
@@ -36,33 +34,4 @@ public class JdbcTransactionIntegrationTest extends DistributedTransactionIntegr
   @Disabled
   @Override
   public void abort_forOngoingTransaction_ShouldAbortCorrectly() {}
-
-  @Disabled("There is no metadata columns when using the jdbc transaction manager")
-  @Override
-  @Test
-  public void scan_UsingDebugMode_ShouldReturnTransactionMetadataColumns() {}
-
-  @Disabled("There is no metadata columns when using the jdbc transaction manager")
-  @Override
-  @Test
-  public void scan_UsingDebugModeWithProjections_ShouldReturnProjectedColumns()
-      throws TransactionException {
-    super.scan_UsingDebugModeWithProjections_ShouldReturnProjectedColumns();
-  }
-
-  @Disabled("There is no metadata columns when using the jdbc transaction manager")
-  @Override
-  @Test
-  public void get_UsingDebugMode_ShouldReturnTransactionMetadataColumns()
-      throws TransactionException {
-    super.get_UsingDebugMode_ShouldReturnTransactionMetadataColumns();
-  }
-
-  @Disabled("There is no metadata columns when using the jdbc transaction manager")
-  @Override
-  @Test
-  public void get_UsingDebugModeWithProjections_ShouldReturnProjectedColumns()
-      throws TransactionException {
-    super.get_UsingDebugModeWithProjections_ShouldReturnProjectedColumns();
-  }
 }
